feat: Update Account DTO and Mapper to include accountTypeId, modify AccountTypeEntity ID type

+++ src/main/java/com/matjazt/netmon2/dto/request/SaveAccountRequest.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 /** Request DTO for creating or updating an Account. */
 public record SaveAccountRequest(
         String username,
-        Long accountTypeId,
+        Integer accountTypeId,
         String password,
         String fullName,
         String email,
